+++ database/views.py
@@ -1,8 +1,6 @@
 from database.models.kinetic_data import BaseKineticsData
 from itertools import zip_longest
 
-import django_filters
-from django.http import HttpResponseRedirect, HttpResponse
 from django.urls import reverse
 from django.core.paginator import Paginator, PageNotAnInteger, EmptyPage
 from django.views import View
@@ -21,6 +19,7 @@ from .models import (
     Reaction,
     Kinetics,
 )
+from .filters import SpeciesFilter, ReactionFilter, SourceFilter
 
 
 class BaseView(TemplateView):
@@ -31,30 +30,7 @@ class IndexView(TemplateView):
     template_name = "index.html"
 
 
-class SpeciesFilter(django_filters.FilterSet):
-    speciesname__name = django_filters.CharFilter(
-        field_name="speciesname", lookup_expr="name", label="Species Name"
-    )
-    isomer__inchi = django_filters.CharFilter(
-        field_name="isomer", lookup_expr="inchi", label="Isomer InChI"
-    )
-    isomer__structure__smiles = django_filters.CharFilter(
-        field_name="isomer", lookup_expr="structure__smiles", label="Structure SMILES"
-    )
-    isomer__structure__adjacency_list = django_filters.CharFilter(
-        field_name="isomer",
-        lookup_expr="structure__adjacency_list",
-        label="Structure Adjacency List",
-    )
-    isomer__structure__electronic_state = django_filters.NumberFilter(
-        field_name="isomer",
-        lookup_expr="structure__electronic_state",
-        label="Structure Electronic State",
-    )
 
-    class Meta:
-        model = Species
-        fields = ["prime_id", "formula", "inchi", "cas_number"]
 
 
 class SpeciesFilterView(FilterView):
@@ -71,37 +47,9 @@ class SpeciesFilterView(FilterView):
             return super_response
 
 
-class SourceFilter(django_filters.FilterSet):
-    sourcename_name = django_filters.CharFilter(
-        field_name="sourcename", lookup_expr="name", label="Source Name"
-    )
-
-    class Meta:
-        model = Source
-        fields = ["name", "prime_id", "publication_year", "source_title", "doi"]
-
-
 class SourceFilterView(FilterView):
     filterset_class = SourceFilter
     paginate_by = 25
-
-    def get(self, request, *args, **kwargs):
-        super_response = super().get(request, *args, **kwargs)
-        pk = request.GET.get("pk")
-        if pk is not None:
-            return HttpResponseRedirect(reverse("source-detail", args=[pk]))
-        else:
-            return super_response
-
-
-class ReactionFilter(django_filters.FilterSet):
-    species__name = django_filters.CharFilter(
-        field_name="species", lookup_expr="speciesname__name", label="Species Name"
-    )
-
-    class Meta:
-        model = Reaction
-        fields = ["prime_id", "reversible"]
 
 
 class ReactionFilterView(FilterView):
